Feat/refactor

+++ src/contracts/strategy/IStrategy.sol
@@ -3,5 +3,5 @@ pragma solidity 0.8.17;
 
 /// Strategy Interface
 interface IStrategy {
-    function getComponentAmount(address coin) external view returns (uint);
+    function getComponentAmount(address coin) external view returns (uint256);
 }
